refactor: extract buildApp into its own file
