+++ package.json
@@ -21,10 +21,10 @@
     "clean": "rimraf dist tsconfig.tsbuildinfo",
     "format": "prettier --write \"{src,test}/**/*.ts\"",
     "build": "tsc -p tsconfig.json && cpy 'src/lib/sql/*.sql' dist/lib/sql",
-    "docs:export": "node --loader ts-node/esm src/server/app.ts docs export > openapi.json",
-    "gen:types:typescript": "node --loader ts-node/esm src/server/app.ts gen types typescript",
-    "start": "node dist/server/app.js",
-    "dev": "trap 'npm run db:clean' INT && run-s db:clean db:run && nodemon --exec node --loader ts-node/esm src/server/app.ts | pino-pretty --colorize",
+    "docs:export": "node --loader ts-node/esm src/server/server.ts docs export > openapi.json",
+    "gen:types:typescript": "node --loader ts-node/esm src/server/server.ts gen types typescript",
+    "start": "node dist/server/server.js",
+    "dev": "trap 'npm run db:clean' INT && run-s db:clean db:run && nodemon --exec node --loader ts-node/esm src/server/server.ts | pino-pretty --colorize",
     "pkg": "run-s clean build && pkg .pkg.config.json",
     "test": "run-s db:clean db:run test:run db:clean",
     "db:clean": "cd test/db && docker-compose down",
